Add -Wconversion to our builds
This commit enables further compiler warnings and adjusts the code
triggering the warnings. In some cases, we can work around the warnings
by tightening input/output types in function signatures. In other cases,
we explain in a comment why the truncation does not alter the value.
Finally, in some cases, the truncation can actually alter the value,
and we rely on signed->unsigned and unsigned->signed integer conversions.

Note that CBMC already checks that integer conversion does not alter
values, so the comments are a helpful indicator that we thought about
this, but not critical.

Resolves https://github.com/pq-code-package/mlkem-native/issues/1229

+++ examples/bring_your_own_fips202/custom_fips202/tiny_sha3/sha3.c
@@ -96,14 +96,14 @@ void sha3_keccakf(uint64_t st[25])
   {
     v = (uint8_t *)&st[i];
     t = st[i];
-    v[0] = t & 0xFF;
-    v[1] = (t >> 8) & 0xFF;
-    v[2] = (t >> 16) & 0xFF;
-    v[3] = (t >> 24) & 0xFF;
-    v[4] = (t >> 32) & 0xFF;
-    v[5] = (t >> 40) & 0xFF;
-    v[6] = (t >> 48) & 0xFF;
-    v[7] = (t >> 56) & 0xFF;
+    v[0] = (uint8_t)(t & 0xFF);
+    v[1] = (uint8_t)((t >> 8) & 0xFF);
+    v[2] = (uint8_t)((t >> 16) & 0xFF);
+    v[3] = (uint8_t)((t >> 24) & 0xFF);
+    v[4] = (uint8_t)((t >> 32) & 0xFF);
+    v[5] = (uint8_t)((t >> 40) & 0xFF);
+    v[6] = (uint8_t)((t >> 48) & 0xFF);
+    v[7] = (uint8_t)((t >> 56) & 0xFF);
   }
 #endif /* __BYTE_ORDER__ != __ORDER_LITTLE_ENDIAN__ */
 }
